refactor(lib): change project structure
