+++ lib/blocs.dart
@@ -1,0 +1,5 @@
+library blocs;
+
+export 'src/blocs/query_wizard_event.dart';
+export 'src/blocs/query_wizard_state.dart';
+export 'src/blocs/query_wizard_bloc.dart';
