+++ lib/infrastructure.dart
@@ -1,0 +1,3 @@
+library infrastructure;
+
+export 'src/infrastructure/query_wizard_bloc_observer.dart';
