+++ lib/layout.dart
@@ -1,0 +1,3 @@
+library layout;
+
+export 'src/layout/adaptive.dart';
