+++ lib/main.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import 'package:query_wizard/src/repositories/repositories.dart';
-import 'package:query_wizard/src/query_wizard_bloc_observer.dart';
-import 'package:query_wizard/src/widgets/query_wizard.dart';
+import 'package:query_wizard/infrastructure.dart';
+import 'package:query_wizard/repositories.dart';
+import 'package:query_wizard/widgets.dart';
 
 void main() {
   Bloc.observer = QueryWizardBlocObserver();
