+++ lib/models.dart
@@ -1,0 +1,3 @@
+library models;
+
+export 'src/models/db_element.dart';
