+++ lib/repositories.dart
@@ -1,0 +1,4 @@
+library repositories;
+
+export 'src/repositories/query_wizard_api_client.dart';
+export 'src/repositories/query_wizard_repository.dart';
