+++ lib/src/infrastructure/query_wizard_bloc_observer.dart
