--- lib/src/repositories/repositories.dart
@@ -1,2 +1,0 @@
-export 'query_wizard_api_client.dart';
-export 'query_wizard_repository.dart';
