+++ lib/src/widgets/query_wizard.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_gen/gen_l10n/query_wizard_localizations.dart';
 
-import 'package:query_wizard/src/blocs/blocs.dart';
-import 'package:query_wizard/src/repositories/repositories.dart';
-import 'package:query_wizard/src/widgets/tabs/tables_and_fields.dart';
+import 'package:query_wizard/blocs.dart';
+import 'package:query_wizard/repositories.dart';
+import 'package:query_wizard/widgets.dart';
 
 class QueryWizard extends StatelessWidget {
   final QueryWizardRepository queryWizardRepository;
