+++ lib/src/widgets/tabs/list_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'package:query_wizard/src/layout/adaptive.dart';
-import 'package:query_wizard/src/models/models.dart';
+import 'package:query_wizard/layout.dart';
+import 'package:query_wizard/models.dart';
 
 class TreeItem extends StatelessWidget {
   const TreeItem({Key key, this.item, this.iconData}) : super(key: key);
