+++ lib/widgets.dart
@@ -1,0 +1,5 @@
+library widgets;
+
+export 'src/widgets/query_wizard.dart';
+export 'src/widgets/tabs/list_item.dart';
+export 'src/widgets/tabs/tables_and_fields.dart';
